+++ packages/components/psammead-timestamp/CHANGELOG.md
@@ -3,5 +3,6 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
-| 0.2.0 | [PR#461](https://github.com/bbc/psammead/pull/461) Make timestamp typography styling agnostic |
+| 0.2.1   | [PR#???](https://github.com/bbc/psammead/pull/???) Add optional padding prop |
+| 0.2.0   | [PR#461](https://github.com/bbc/psammead/pull/461) Make timestamp typography styling agnostic |
 | 0.1.0   | [PR#440](https://github.com/bbc/psammead/pull/440) Create initial package with Timestamp component pulled in from [Simorgh](https://github.com/BBC-News/simorgh). |